+++ KruacentE.GlobalEventFramework/GEFE.Examples/GE/R.cs
@@ -14,7 +14,7 @@ namespace GEFExiled.GEFE.Examples.GE
         public override int Id { get; set; } = 32;
         public override string Name { get; set; } = "nothing";
         public override string Description { get; set; } = "y'a r";
-        public override double Weight { get; set; } = 1;
+        public override int Weight { get; set; } = 1;
         public override IEnumerator<float> Start()
         {
             yield return 0;
